+++ .snapshots/successful/adt_enum/compiler/compiled.js
@@ -49,7 +49,7 @@ function v_show(v_c){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_18_9 = s[1]; return [3, v__do_e_18_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_19_9 = s[1]; return [3, v__do_e_19_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_20_9 = s[1]; return [3, v__do_e_20_9]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, (v_show)([22])); } } })(__concat(v_s1, ", ")); } } })(__concat(v_s0, (v_show)([21]))); } } })(__concat((v_show)([20]), ", ")));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_19_9 = s[1]; return [3, v__do_e_19_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_20_9 = s[1]; return [3, v__do_e_20_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_21_9 = s[1]; return [3, v__do_e_21_9]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, (v_show)([22])); } } })(__concat(v_s1, ", ")); } } })(__concat(v_s0, (v_show)([21]))); } } })(__concat((v_show)([20]), ", ")));
 
 function v__let_12(v_res){
     {