+++ .snapshots/successful/prelude_bool-ops/compiler/compiled.js
@@ -88,7 +88,7 @@ function v_runIO(v_io){
   }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_7_9 = s[1]; return [3, v__do_e_7_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_8_9 = s[1]; return [3, v__do_e_8_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_9_9 = s[1]; return [3, v__do_e_9_9]; } case 4: { const v_s2 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_10_9 = s[1]; return [3, v__do_e_10_9]; } case 4: { const v_s3 = s[1]; return __concat(v_s3, (v_showBool)((v_or)([1], [2]))); } } })(__concat(v_s2, (v_showBool)((v_or)([2], [2])))); } } })(__concat(v_s1, (v_showBool)((v_and)([1], [1])))); } } })(__concat(v_s0, (v_showBool)((v_and)([1], [2])))); } } })(__concat((v_showBool)((v_not)([1])), (v_showBool)((v_not)([2])))));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_8_9 = s[1]; return [3, v__do_e_8_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_9_9 = s[1]; return [3, v__do_e_9_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_10_9 = s[1]; return [3, v__do_e_10_9]; } case 4: { const v_s2 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_11_9 = s[1]; return [3, v__do_e_11_9]; } case 4: { const v_s3 = s[1]; return __concat(v_s3, (v_showBool)((v_or)([1], [2]))); } } })(__concat(v_s2, (v_showBool)((v_or)([2], [2])))); } } })(__concat(v_s1, (v_showBool)((v_and)([1], [1])))); } } })(__concat(v_s0, (v_showBool)((v_and)([1], [2])))); } } })(__concat((v_showBool)((v_not)([1])), (v_showBool)((v_not)([2])))));
 
 function v__let_12(v_res){
     {