+++ .snapshots/successful/prelude_either/compiler/compiled.js
@@ -48,7 +48,7 @@ function v_unwrap(v_r){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_12_9 = s[1]; return [3, v__do_e_12_9]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, v_b); } } })(__concat(v_a, ", ")); } } })((v_unwrap)([4, "good"])); } } })((v_unwrap)([3, "bad"])));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_15_9 = s[1]; return [3, v__do_e_15_9]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, v_b); } } })(__concat(v_a, ", ")); } } })((v_unwrap)([4, "good"])); } } })((v_unwrap)([3, "bad"])));
 
 function v__let_12(v_res){
     {