+++ .snapshots/successful/prelude_maybe/compiler/compiled.js
@@ -47,7 +47,7 @@ function v_unwrap(v_m){
     }
 }
 
-const main = (v__let_13)(((s) => { switch(s[0]) { case 3: { const v__do_e_12_9 = s[1]; return [3, v__do_e_12_9]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, v_b); } } })(__concat(v_a, ", ")); } } })((v_unwrap)((v__lift_12)([10]))); } } })((v_unwrap)([11, "hi"])));
+const main = (v__let_13)(((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_15_9 = s[1]; return [3, v__do_e_15_9]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, v_b); } } })(__concat(v_a, ", ")); } } })((v_unwrap)((v__lift_12)([10]))); } } })((v_unwrap)([11, "hi"])));
 
 function v__lift_12(v___input){
     {