+++ .snapshots/successful/smoke_hello/compiler/compiled.js
@@ -90,8 +90,8 @@ function v_addGreeting(v_name){
       const __s = __concat(v_greeting, ", ");
       switch (__s[0]) {
         case 3: {
-          const v__do_e_27_3 = __s[1];
-          return [3, v__do_e_27_3];
+          const v__do_e_28_3 = __s[1];
+          return [3, v__do_e_28_3];
         }
         case 4: {
           const v_s0 = __s[1];
@@ -99,8 +99,8 @@ function v_addGreeting(v_name){
             const __s = __concat(v_s0, v_name);
             switch (__s[0]) {
               case 3: {
-                const v__do_e_28_3 = __s[1];
-                return [3, v__do_e_28_3];
+                const v__do_e_29_3 = __s[1];
+                return [3, v__do_e_29_3];
               }
               case 4: {
                 const v_s1 = __s[1];