+++ .snapshots/successful/syntax_do-let-binding/compiler/compiled.js
@@ -71,8 +71,8 @@ function v_run(v_start){
       const __s = (v_step1)(v_start);
       switch (__s[0]) {
         case 3: {
-          const v__do_e_18_3 = __s[1];
-          return [3, [1615808600, v__do_e_18_3]];
+          const v__do_e_19_3 = __s[1];
+          return [3, [1615808600, v__do_e_19_3]];
         }
         case 4: {
           const v_a = __s[1];
@@ -112,8 +112,8 @@ function v__let_12(v_a, v_prefix){
       const __s = (v_step2)(v_a);
       switch (__s[0]) {
         case 3: {
-          const v__do_e_20_3 = __s[1];
-          return [3, [1615808600, v__do_e_20_3]];
+          const v__do_e_21_3 = __s[1];
+          return [3, [1615808600, v__do_e_21_3]];
         }
         case 4: {
           const v_b = __s[1];