+++ .snapshots/successful/type-ascription-expression/compiler/compiled.js
@@ -43,7 +43,7 @@ const v_throughHof = String((v_identity)((10|0)));
 
 const v_ascribedCall = ((s) => { switch(s[0]) { case 3: { const v___w0 = s[1]; return "FAIL"; } case 4: { const v_n = s[1]; return String(v_n); } } })(__predInt32((5|0)));
 
-const v_joined = ((s) => { switch(s[0]) { case 3: { const v__do_e_34_3 = s[1]; return [3, v__do_e_34_3]; } case 4: { const v_ab = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_35_3 = s[1]; return [3, v__do_e_35_3]; } case 4: { const v_abc = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_36_3 = s[1]; return [3, v__do_e_36_3]; } case 4: { const v_abcd = s[1]; return __concat(v_abcd, v_ascribedCall); } } })(__concat(v_abc, " ")); } } })(__concat(v_ab, v_throughHof)); } } })(__concat(v_direct, " "));
+const v_joined = ((s) => { switch(s[0]) { case 3: { const v__do_e_35_3 = s[1]; return [3, v__do_e_35_3]; } case 4: { const v_ab = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_36_3 = s[1]; return [3, v__do_e_36_3]; } case 4: { const v_abc = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_37_3 = s[1]; return [3, v__do_e_37_3]; } case 4: { const v_abcd = s[1]; return __concat(v_abcd, v_ascribedCall); } } })(__concat(v_abc, " ")); } } })(__concat(v_ab, v_throughHof)); } } })(__concat(v_direct, " "));
 
 const main = ((s) => { switch(s[0]) { case 3: { const v___w0 = s[1]; return [7, "OVERFLOW", [5, [0]]]; } case 4: { const v_s = s[1]; return [7, v_s, [5, [0]]]; } } })(v_joined);