+++ .snapshots/successful/uint32_eq/compiler/compiled.js
@@ -51,7 +51,7 @@ function v_render(v_b){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_12_9 = s[1]; return [3, v__do_e_12_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, (v_render)(__eqUInt32((2147483648 >>> 0), (2147483647 >>> 0)))); } } })(__concat(v_s1, (v_render)(__eqUInt32((2147483648 >>> 0), (2147483648 >>> 0))))); } } })(__concat(v_s0, (v_render)(__eqUInt32(v_maxUInt32, v_minUInt32)))); } } })(__concat((v_render)(__eqUInt32(v_minUInt32, v_minUInt32)), (v_render)(__eqUInt32(v_maxUInt32, v_maxUInt32)))));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_15_9 = s[1]; return [3, v__do_e_15_9]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, (v_render)(__eqUInt32((2147483648 >>> 0), (2147483647 >>> 0)))); } } })(__concat(v_s1, (v_render)(__eqUInt32((2147483648 >>> 0), (2147483648 >>> 0))))); } } })(__concat(v_s0, (v_render)(__eqUInt32(v_maxUInt32, v_minUInt32)))); } } })(__concat((v_render)(__eqUInt32(v_minUInt32, v_minUInt32)), (v_render)(__eqUInt32(v_maxUInt32, v_maxUInt32)))));
 
 function v__let_12(v_res){
     {