+++ .snapshots/successful/uint8_eq/compiler/compiled.js
@@ -51,7 +51,7 @@ function v_render(v_b){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_12_9 = s[1]; return [3, v__do_e_12_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_s1 = s[1]; return __concat(v_s1, (v_render)(__eqUInt8((128 & 0xFF), (127 & 0xFF)))); } } })(__concat(v_s0, (v_render)(__eqUInt8(v_maxUInt8, v_minUInt8)))); } } })(__concat((v_render)(__eqUInt8(v_minUInt8, v_minUInt8)), (v_render)(__eqUInt8(v_maxUInt8, v_maxUInt8)))));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_13_9 = s[1]; return [3, v__do_e_13_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14_9 = s[1]; return [3, v__do_e_14_9]; } case 4: { const v_s1 = s[1]; return __concat(v_s1, (v_render)(__eqUInt8((128 & 0xFF), (127 & 0xFF)))); } } })(__concat(v_s0, (v_render)(__eqUInt8(v_maxUInt8, v_minUInt8)))); } } })(__concat((v_render)(__eqUInt8(v_minUInt8, v_minUInt8)), (v_render)(__eqUInt8(v_maxUInt8, v_maxUInt8)))));
 
 function v__let_12(v_res){
     {